# docs/api.yaml
openapi: 3.0.0
info:
  title: Wildfire Decision Support API
  version: 4.0.0
  description: >
    REST API for the Wildfire Decision Support System (ENGO 651).
    All GeoJSON geometries are in EPSG:4326 (WGS84).
    Protected endpoints require a short-lived access JWT in the Bearer header.
    Refresh JWTs are rotated and may only be sent to /auth/refresh or /auth/logout.
servers:
  - url: http://localhost:5000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventTimestep:
      type: object
      properties:
        id:            { type: integer, example: 42 }
        slot_time:     { type: string, format: date-time, example: "2016-05-04T06:00:00" }
        nearest_t1:    { type: string, format: date-time, example: "2016-05-04T05:12:00" }
        gap_hours:     { type: number, example: 0.8 }
        data_gap_warn: { type: boolean, example: false }

    # ── GeoJSON: Perimeter ────────────────────────────────────────────────────
...
paths:
...
  /api/events/{event_id}/timesteps:
    get:
      summary: List all hourly timestep slots for an event
      description: >
        Returns the hourly slot grid from event start to end.
        Each slot has nearest_t1 (most recent satellite overpass ≤ slot_time).
        Pipeline stage statuses are read from STATUS.json files on disk, not stored in DB.
      parameters:
        - { name: event_id, in: path, required: true, schema: { type: integer } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EventTimestep' }
        404: { description: Event not found }
  /api/events/{event_id}/timesteps/{ts_id}/status:
    get:
      summary: Get pipeline status for a specific timestep
      description: Reads STATUS.json files from prediction/ML/ and spatial_analysis/ML/ directories.
      parameters:
        - { name: event_id, in: path, required: true, schema: { type: integer } }
        - { name: ts_id,    in: path, required: true, schema: { type: integer } }
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction_status:       { type: string, enum: [pending, running, done, failed] }
                  spatial_analysis_status: { type: string, enum: [pending, running, done, failed] }
        404: { description: Event or timestep not found }
  /api/events/{event_id}/timesteps/{ts_id}/run-prediction:
    post:
      summary: Trigger prediction pipeline for a timestep (DEV / admin only)
      description: >
        Runs the full pipeline: environment data → ML prediction → spatial analysis.
        Non-blocking — returns immediately. Poll /status to track progress.
      parameters:
        - { name: event_id, in: path, required: true, schema: { type: integer } }
        - { name: ts_id,    in: path, required: true, schema: { type: integer } }
      responses:
        202: { description: Pipeline started }
        404: { description: Event or timestep not found }
...
  /api/events/{event_id}/chat:
    post:
      summary: Stateless streaming chat with wildfire decision support assistant
      description: >
        Streams plain text chunks (text/plain).
        When timestep_id is provided, the agent is seeded with the AI_report context
        (situation briefing, risk level, key points, and all specialist analysis).
        Each response ends with a "Suggested questions:" section listing 3 follow-up questions
        that are directly answerable from the loaded report.
      parameters:
        - { name: event_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: "What areas should be evacuated first?"
                timestep_id:
                  type: integer
                  nullable: true
                  description: Seeds chat context with AI_report for this timestep
                history:
                  type: array
                  description: Prior turns for multi-turn conversation
                  items:
                    type: object
                    properties:
                      role:    { type: string, enum: [user, assistant] }
                      content: { type: string }
      responses:
        200:
          description: Streamed plain-text response
          content:
            text/plain:
              schema: { type: string }
        400: { description: message required }

  # ── Field Reports (Crowd Intelligence) ───────────────────────────────────
...